+++ Makefile
@@ -30,19 +30,13 @@ virtualenv-dev: setup-virtualenv
 # ====
 
 lint: virtualenv-dev
-	$(flakehell) lint data_generator query_timer
-	$(flakehell) lint batch_size_automator float_simulator tictrack
-	$(flakehell) lint tests
+	$(flakehell) lint tsdg tests
 
 format: virtualenv-dev
-	$(black) data_generator query_timer
-	$(black) batch_size_automator float_simulator tictrack
-	$(black) tests
+	$(black) tsdg tests
 
 test: virtualenv-dev
 	$(pytest) -vvv tests
 
 coverage: virtualenv-dev
-	$(pytest) \
-	    --cov=data_generator --cov=query_timer --cov=batch_size_automator --cov=float_simulator --cov=tictrack \
-	    --cov-fail-under=95 --cov-branch --cov-report=term tests
+	$(pytest) --cov=tsdg --cov-fail-under=95 --cov-branch --cov-report=term tests
