+++ tests/util/test_batch_size_automator.py
@@ -1,4 +1,4 @@
-from batch_size_automator import BatchSizeAutomator
+from tsdg.util.batch_size_automator import BatchSizeAutomator
 
 
 def test_is_auto_batch_mode():
