+++ tsdg/adapter/influxdb.py
@@ -21,8 +21,8 @@
 
 from typing import Tuple
 
-from tictrack import timed_function
-from data_generator.db_writer import DbWriter
+from tsdg.util.tictrack import timed_function
+from tsdg.model.database import DbWriter
 from influxdb_client import InfluxDBClient, Bucket
 from influxdb_client.client.write_api import SYNCHRONOUS, Point
 from datetime import datetime
