+++ tsdg/adapter/mongodb.py
@@ -21,8 +21,8 @@
 
 from typing import Tuple
 
-from tictrack import timed_function
-from data_generator.db_writer import DbWriter
+from tsdg.util.tictrack import timed_function
+from tsdg.model.database import DbWriter
 from pymongo import MongoClient
 from pymongo import CursorType
 from datetime import datetime
