+++ tsdg/adapter/postgresql.py
@@ -21,30 +21,27 @@
 
 import psycopg2
 import psycopg2.extras
-from pgcopy import CopyManager
-from tictrack import timed_function
-from data_generator.db_writer import DbWriter
+from tsdg.util.tictrack import timed_function
+from tsdg.model.database import DbWriter
 from datetime import datetime
 from datetime_truncate import truncate
 
 
-class TimescaleDbWriter(DbWriter):
+class PostgresDbWriter(DbWriter):
     def __init__(
         self,
         host: str,
         port: int,
         username: str,
         password: str,
-        ts_db_name: str,
+        db_name: str,
         model: dict,
         table_name: str = None,
         partition: str = "week",
-        copy: bool = False,
-        distributed: bool = False,
     ):
         super().__init__()
         self.conn = psycopg2.connect(
-            dbname=ts_db_name, user=username, password=password, host=host, port=port
+            dbname=db_name, user=username, password=password, host=host, port=port
         )
         self.cursor = self.conn.cursor()
         self.model = model
@@ -52,8 +49,6 @@ class TimescaleDbWriter(DbWriter):
             table_name is None or table_name == ""
         ]
         self.partition = partition
-        self.copy = copy
-        self.distributed = distributed
 
     def close_connection(self):
         self.cursor.close()
@@ -71,47 +66,15 @@ ts_{self.partition} TIMESTAMP NOT NULL,
 
         self.cursor.execute(stmt)
         self.conn.commit()
-        if self.distributed:
-            tag = self._get_partition_tag()
-            stmt = f"SELECT create_distributed_hypertable('{self.table_name}', 'ts', '{tag}', if_not_exists => true);"
-        else:
-            stmt = (
-                f"SELECT create_hypertable('{self.table_name}', 'ts', 'ts_{self.partition}', 10, "
-                f"if_not_exists => true);"
-            )
+
+    @timed_function()
+    def insert_stmt(self, timestamps: list, batch: list):
+        stmt = self._prepare_postgres_stmt(timestamps, batch)
         self.cursor.execute(stmt)
         self.conn.commit()
 
     @timed_function()
-    def insert_stmt(self, timestamps: list, batch: list):
-        if self.copy:
-            self._prepare_copy(timestamps, batch)
-        else:
-            stmt = self._prepare_timescale_stmt(timestamps, batch)
-            self.cursor.execute(stmt)
-        self.conn.commit()
-
-    @timed_function()
-    def _prepare_copy(self, timestamps: list, batch: list):
-        columns = self._get_tags_and_metrics().keys()
-        values = []
-
-        for i in range(0, len(timestamps)):
-            t = datetime.fromtimestamp(timestamps[i] / 1000)
-            trunc = truncate(t, self.partition)
-            data = [t, trunc]
-            for column in columns:
-                data.append(batch[i][column])
-            values.append(data)
-
-        cols = ["ts", f"ts_{self.partition}"]
-        for column in columns:
-            cols.append(column)
-        copy_manager = CopyManager(self.conn, self.table_name, cols)
-        copy_manager.copy(values)
-
-    @timed_function()
-    def _prepare_timescale_stmt(self, timestamps: list, batch: list) -> str:
+    def _prepare_postgres_stmt(self, timestamps: list, batch: list) -> str:
         columns = self._get_tags_and_metrics().keys()
         stmt = f"""INSERT INTO {self.table_name} (ts, ts_{self.partition},"""
         for column in columns:
@@ -137,10 +100,3 @@ ts_{self.partition} TIMESTAMP NOT NULL,
         for key in self.model.keys():
             if key != "description":
                 return key
-
-    def _get_partition_tag(self, top_level: bool = False) -> str:
-        key = self._get_model_table_name()
-        tags = list(self.model[key]["tags"].keys())
-        if "description" in tags:
-            tags.remove("description")
-        return tags[0] if top_level else tags[-1]
