+++ tsdg/util/batch_size_automator/__init__.py
