+++ tsdg/util/tictrack/README.md
@@ -34,17 +34,19 @@ functions the wrapper funciton is the better solution.
 This is the easiest way to use `tictrack`:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 
 @tictrack.timed_function()
 def foo(a, b):
-    return a + b
+  return a + b
 
 
 foo_results = []
 for i in range(0, 10):
-    for j in range(10, 20):
-        foo_results = foo(i, j)
+  for j in range(10, 20):
+    foo_results = foo(i, j)
 ```
 
 You can than calculate the average execution time like so:
@@ -58,20 +60,23 @@ The decorator supports any type of function with any combination of arguments, *
 You can also track nested functions like this:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 
 @tictrack.timed_function()
 def foo(a, b):
-    return a + b
+  return a + b
+
 
 @tictrack.timed_function()
 def bar(x):
-    return foo(x, x)
+  return foo(x, x)
 
 
 bar_results = []
 for i in range(0, 10):
-    bar_results = bar(i)
+  bar_results = bar(i)
 ```
 
 You can now calculate the average execution time of `foo` and `bar` like so:
@@ -97,10 +102,12 @@ could set `do_print=True` and `save_result=False`.
 The Wrapper can be used to only track certain function executions:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 
 def foo(a, b):
-    return a + b
+  return a + b
 
 
 x = foo(1, 2)  # this is not tracked
@@ -141,26 +148,29 @@ average execution time of foo. Additionally, the input parameters for `foo` chan
 track for the averages each minute as well:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 
 @tictrack.timed_function()
 def foo(a, b):
-    # code
-    pass
+  # code
+  pass
+
 
 delta_averages = []
 
 while not ten_minutes_have_passed:
-    ...
-    foo(a, b)
-    if one_minute_has_passed:
-        delta_averages.append(tictrack.timed_function_statistics("foo", delta=True))
-        tictrack.reset_delta("foo")
-    ...
+  ...
+  foo(a, b)
+  if one_minute_has_passed:
+    delta_averages.append(tictrack.timed_function_statistics("foo", delta=True))
+    tictrack.reset_delta("foo")
+  ...
 
 # print the results
 for avg in delta_averages:
-    print(f"foo delta took {avg} seconds on average")
+  print(f"foo delta took {avg} seconds on average")
 
 print(f"foo took {tictrack.timed_function_statistics('foo')} seconds on average")
 
@@ -171,7 +181,9 @@ print(f"foo took {tictrack.timed_function_statistics('foo')} seconds on average"
 The result set generated by `tictrack` can be analyzed by directly accessing the result set like so:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 ...
 foo_result = tictrack.tic_toc["foo"]  # general results
 foo_delta_result = tictrack.tic_toc_delta["foo"]  # delta results
@@ -207,11 +219,14 @@ function can be used. It applies the given function to a result set and saves th
 by default `statistics.mean` is used:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
+
 
 @tictrack.timed_function()
 def foo(a, b):
-    pass
+  pass
+
 
 # code
 # tictrack.tic_toc["foo"] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
@@ -233,7 +248,8 @@ It is also possible to delete an entire result set by calling either `tictrack.r
 `tictrack.reset_delta("function_name")`:
 
 ```python
-import tictrack
+
+from tsdg.util import tictrack
 
 # tictrack.tic_toc["foo"] = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9]
 tictrack.reset("foo")
